refactor(LottoRule): 검증을 입력에서만 받기 때문에 굳이 클래스 분리를 할 필요가 없어서 제거. 하지만 값들은 ErrorMessage때문에 필요하기 때문에 룰로 바뀔수 있는 값들을 LottoRule이라는 enum 클래스로 따로 묶어서 저장
