+++ src/main/java/lotto/validation/ErrorMessage.java
@@ -7,7 +7,7 @@ public enum ErrorMessage {
     PURCHASE_AMOUNT_POSITIVE(() -> "구입 금액은 양수여야 합니다."),
     PURCHASE_AMOUNT_ZERO(() -> "구입 금액은 0원일 수 없습니다."),
     PURCHASE_AMOUNT_MAX(() -> "최대 구입 금액을 초과했습니다."),
-    WINNING_NUMBER_RANGE(() -> String.format("당첨 번호는 %d부터 %d 사이의 정수여야 합니다.", WinningNumberValidation.MIN_NUMBER, WinningNumberValidation.MAX_NUMBER)),
+    WINNING_NUMBER_RANGE(() -> String.format("당첨 번호는 %d부터 %d 사이의 정수여야 합니다.", LottoRule.MIN_NUMBER.getValue(), LottoRule.MAX_NUMBER.getValue())),
     WINNING_NUMBER_COUNT(() -> "당첨 번호는 6개여야 합니다."),
     DUPLICATE_WINNING_NUMBER(() -> "당첨 번호는 중복될 수 없습니다."),
     WINNING_NUMBER_FORMAT(() -> "당첨 번호는 쉼표(,)로 구분된 숫자여야 합니다."),
